requirements.txt
aiosqlite==0.17.0
anyio==3.6.2
asyncpg==0.27.0
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.95.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
orjson==3.8.10
passlib==1.7.4
psycopg==3.0.12
psycopg-binary==3.0.12
psycopg-pool==3.1.7
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
pypika-tortoise==0.1.6
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
tortoise-orm==0.19.3
typing_extensions==4.5.0
ujson==5.7.0
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.1
